% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay.R
\name{add_turn_info}
\alias{add_turn_info}
\title{Add turn information}
\usage{
add_turn_info(replay)
}
\arguments{
\item{replay}{a \code{bb2_replay} object}
}
\value{
the passed in object with turn information added
}
\description{
Summarises the active team, game phase and game turn for each step of a replay. Useful if you then want to run analysis on only a per-player-turn basis
}
\examples{
r = read_replay(test_file()) \%>\% add_turn_info()
}
