% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_call}
\alias{api_call}
\title{Generic API call method}
\usage{
api_call(method, params, simplify = F, debug = F)
}
\arguments{
\item{method}{The name of the method to call}

\item{params}{The query parameters to call the method with}

\item{simplify}{Should the resulting data be simplified? \code{\link[jsonlite](fromJSON)} parameter \code{simplifyDataFrame}.}
}
\value{
The data from the specified API call as a list
}
\description{
To be customised for each available method
}
