% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_coaches}
\alias{api_coaches}
\title{Coach information}
\usage{
api_coaches(key = NA, league = NA, competition = NA, platform = NA,
  limit = NA, ...)
}
\arguments{
\item{key}{API access key (required)}

\item{league}{League name (defaults to Cabalvision Official League)}

\item{competition}{Competition name (defaults to Free Ladder)}

\item{platform}{pc, ps4, or xb1 (optional)}

\item{limit}{Number of entries to return (defaults to 100)}
}
\value{
List containing two named elements:
\itemize{
\item{\code{coaches}: a data.frame of coach information - name, id, country, language}
\item{\code{meta}: a list containing information about the query}
}
}
\description{
Returns information on coaches for a specified league/competition
}
\examples{

}
