% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_halloffame}
\alias{api_halloffame}
\title{Hall of Fame stats for a league/competition}
\usage{
api_halloffame(key = NULL, league = NULL, competition = NULL,
  platform = NULL, limit = NULL, exact = 1)
}
\arguments{
\item{key}{API access key (required)}

\item{league}{League name (defaults to Cabalvision Official League)}

\item{competition}{Competition name (defaults to all competitions in league)}

\item{platform}{pc, ps4, or xb1 (optional)}

\item{limit}{Number of statistics to return (defaults to 100)}

\item{exact}{0 or 1 - Exact matching for league name}
}
\description{
Hall of Fame stats for a league/competition
}
