% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_ladder}
\alias{api_ladder}
\title{Competition Ladder}
\usage{
api_ladder(key = NA, league = NA, competition = NA,
  ladder_size = NA, ...)
}
\arguments{
\item{key}{API access key (required)}

\item{league}{League name (defaults to Cabalvision Official League)}

\item{competition}{Competition name (defaults to Free Ladder)}

\item{ladder_size}{Size of ladder to return (defaults to 100)}
}
\value{
List containing three named elements:
\itemize{
\item{\code{ranking}: a data.frame of the ladder ranking - team.rank, team.name, team.id, team.tv, team.score, team.race, team.logo, coach.name, coach.id, coach.lang, coach.country}
\item{\code{ladder}: a list of information about the ladder - league/competition names, ladder id}
\item{\code{meta}: a list containing information about the query}
}
}
\description{
Competition Ladder
}
