% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_league}
\alias{api_league}
\title{League information}
\usage{
api_league(key = NA, league = NA, platform = NA, ...)
}
\arguments{
\item{key}{API access key (required)}

\item{league}{League name (defaults to Cabalvision Official League)}

\item{platform}{pc, ps4, or xb1 (optional)}
}
\value{
List containing two named elements:
\itemize{
\item{\code{league}: a list of league information - date_last_match, logo, treasury, team_count}
\item{\code{meta}: a list containing information about the query}
}
}
\description{
Returns information about a specified league
}
