% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_match}
\alias{api_match}
\title{Detailed match information}
\usage{
api_match(key = NULL, match_id = NULL, platform = NULL, ...)
}
\arguments{
\item{key}{API access key (required)}

\item{match_id}{UUID for the match (required)}

\item{platform}{pc, ps4, or xb1 (optional)}
}
\description{
Returns detailed match information from a BB2 match
}
\examples{

}
