% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_matches}
\alias{api_matches}
\title{List matches for a league/competition}
\usage{
api_matches(key = NA, league = NA, competition = NA, platform = NA,
  limit = NA, id_only = NA, start = NA, end = NA, ...)
}
\arguments{
\item{key}{API access key (required)}

\item{league}{League name (defaults to Cabalvision Official League)}

\item{competition}{Competition name (optional)}

\item{platform}{pc, ps4, or xb1 (optional)}

\item{limit}{Number of entries to return (defaults to 100)}

\item{id_only}{Return match summaries or just the match id (0|1, defaults to match summary)}

\item{start}{Start date (defaults to 20 days ago)}

\item{end}{End date (defaults to today)}
}
\value{
List containing four named elements:
\itemize{
\item{\code{debug}: a list containing the query parameters used}
\item{\code{matches}: a list containing basic information about each match - uuid, league/competition, coaches, teams and summarised statistics}
\item{\code{urls}: a list containing urls for various game assets}
\item{\code{meta}: a list containing information about the query}
}
}
\description{
Returns information on matches played in the specified league/competition. Starts from the most recent match within the specified window and works backwards.
}
