% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_team}
\alias{api_team}
\title{Team information}
\usage{
api_team(key = NULL, name = NULL, id = NULL, platform = NULL, ...)
}
\arguments{
\item{key}{API access key (required)}

\item{name}{Team name (ignored if id is provided)}

\item{id}{Team id}

\item{platform}{pc, ps4, or xb1 (optional)}
}
\value{
List containing five named elements:
\item{\code{team}: a list of information about the team - name, race, rerolls, apo, nbplayers}
\item{\code{coach}: a list of information about the team's coach}
\item{\code{roster}: a list of information about the players on the team - id, name, value, stats, xp, skills, casualties_state}
\item{\code{urls}: a list of urls for various game assets}
\item{\code{meta}: a list of information about the query}
}
\description{
Team information
}
