% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_teams}
\alias{api_teams}
\title{Teams information}
\usage{
api_teams(key = NA, league = NA, competition = NA, platform = NA,
  limit = NA, ...)
}
\arguments{
\item{key}{API access key (required)}

\item{league}{League name (defaults to Cabalvision Official League)}

\item{competition}{Competition name (optional)}

\item{platform}{pc, ps4, or xb1 (optional)}

\item{limit}{Number of entries to return (defaults to 100)}
}
\value{
List containing three named elements:
\itemize{
\item{\code{teams}: a data.frame of team information - team, id, coach, logo, race_id, race, description (motto), dateLastMatch}
\item{\code{meta}: a list containing information about the query}
\item{\code{urls}: a list of urls for various game assets}
}
}
\description{
Returns information for teams in a league/competition
}
