% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{calc_game_probs}
\alias{calc_game_probs}
\title{Calculate game probabilities}
\usage{
calc_game_probs(home_team, away_team, weights)
}
\arguments{
\item{home_team}{\code{team} object for the home team}

\item{away_team}{\code{team} object for the away team}

\item{weights}{Vector of weights for probabilities. Length three, order is CCL (TV difference), Home team, Away team}
}
\value{
List of probabilities for home team game outcomes
}
\description{
Combine probabilites derived from the home and away team's record, as well as a TV based win probability from CCL data, to determine probability of outcomes for a specific game. Relative importance of the factors is adjusted with the \code{weights} parameter. Function adds a little bit of randomness to the final calculated properties to avoid a deterministic outcome.
}
