% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay_step.R
\name{extract_game_state}
\alias{extract_game_state}
\title{Extract game state}
\usage{
extract_game_state(replay_step)
}
\arguments{
\item{replay_step}{}
}
\description{
Identifies active team and game phase/turn
}
