% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay.R
\name{extract_replay_metadata}
\alias{extract_replay_metadata}
\title{Extract replay metadata}
\usage{
extract_replay_metadata(xml_data, replay_length, metadata_required)
}
\arguments{
\item{xml_data}{an \code{xml_document} resulting from reading in a .bbrz file with \code{xml2::read_xml}}

\item{replay_length}{How many nodes \code{xml_data} contains. The match results are contained in the final entry}
}
\value{
a list of relevant metadata as strings
}
\description{
Extracts the metadata from an unzipped .bbrz file.
}
\details{
Relevant metadata is found in the final xml entry, reporting the game results back to the server.
}
