% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_stars}
\alias{find_stars}
\title{Find Star Players}
\usage{
find_stars(uuid)
}
\arguments{
\item{uuid}{UUID of match to check}
}
\value{
Prints a list of unrecorded star players and the team they were hired for
}
\description{
Utility to find star players to record them properly in mappings.R
}
