% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{flip_y}
\alias{flip_y}
\title{Flip y values}
\usage{
flip_y(y_vals)
}
\arguments{
\item{y_vals}{integers to flip.}
}
\description{
Cyanide's way of recording y values produces a mirror image of what is displayed on the pitch. This function mirrors y values around the pitch midpoint (7) to produce plots that look like what the CabalVision replay looks like
}
