% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay.R
\name{format.bb2_replay}
\alias{format.bb2_replay}
\title{Format a \code{bb2_replay}}
\usage{
\method{format}{bb2_replay}(x, ...)
}
\arguments{
\item{x}{a \code{bb2_replay} object}

\item{...}{}
}
\description{
Format a \code{bb2_replay}
}
