% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_stats.R
\name{get_game_stats}
\alias{get_game_stats}
\title{Get game statistics}
\usage{
get_game_stats(uuid, platform)
}
\arguments{
\item{uuid}{String of the uuid for the game}
}
\value{
A list from the JSON match stats returned from the Cyanide API
}
\description{
Get game statistics
}
