% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_turn.R
\name{get_turn_steps}
\alias{get_turn_steps}
\title{Get replay steps for game turn}
\usage{
get_turn_steps(replay, team = c("home", "away"), turn = 0)
}
\arguments{
\item{replay}{A \code{bb2_replay} object}

\item{team}{Which team to work on. Must be either "home" or "away"}

\item{turn}{Game turn to extract. Must be a valid turn number for the replay}
}
\value{
A vector of replay stepIDs for the team and game turn.
}
\description{
Extracts the replay steps corresponding to a specified team's game turn. \code{add_turn_info()} must have been run on the replay object first.
}
