% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_turn.R
\name{modify_rolled_dice}
\alias{modify_rolled_dice}
\title{Modifies rolled dice types}
\usage{
modify_rolled_dice(type, dice_list)
}
\arguments{
\item{type}{string describing the roll type}

\item{dice_list}{string with the bracketed list of dice rolled}
}
\description{
Removes the doubled up dice for block dice
}
