% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_data}
\alias{parse_data}
\title{Parse arbitrary xml data}
\usage{
parse_data(nodeset, data, data_type, null_val = NA)
}
\arguments{
\item{nodeset}{an xml_nodest object to parse through step by step}

\item{data}{the xpath search string for the data to extract}

\item{data_type}{what format to convert the data into (eg. xml2::xml_integer)}

\item{null_val}{what to record null values as}
}
\description{
Parse arbitrary xml data
}
