% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_stats.R
\name{parse_injuries}
\alias{parse_injuries}
\title{Parse injuries taken in a game}
\usage{
parse_injuries(player_result)
}
\arguments{
\item{player_result}{an element from the playerResults list from Cyanide's GetMatch API for a team}
}
\value{
The injuries sustained by the player during the game
}
\description{
Parse injuries taken in a game
}
