% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_stats.R
\name{parse_perms}
\alias{parse_perms}
\title{Parse permanent injuries for a player}
\usage{
parse_perms(cas_list)
}
\arguments{
\item{cas_list}{listCasualties from the playerResults list from Cyanide's GetMatch API for a team}
}
\value{
Any permanent injuries (niggle or stat bust) sustained by a player
}
\description{
Parse permanent injuries for a player
}
