% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_turn.R
\name{parse_player_turn}
\alias{parse_player_turn}
\title{Parses an entire player turn from a \code{bb2_replay} object}
\usage{
parse_player_turn(replay, team = c("home", "away"), turn = 0)
}
\arguments{
\item{replay}{A \code{bb2_replay} object}

\item{team}{Which team to work on. Must be either "home" or "away"}

\item{turn}{Game turn to extract. Must be a valid turn number for the replay}
}
\value{
A \code{player_turn} object
}
\description{
Parses an entire player turn from a \code{bb2_replay} object
}
