% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_stats.R
\name{player_data}
\alias{player_data}
\title{Get Player Data}
\usage{
player_data(match_data, team)
}
\arguments{
\item{match_data}{Match data object from \link{\code{get_game_stats}}}

\item{team}{One of \code{"home"} or \code{"away"} to identify the team}
}
\value{
A \code{data_frame} containing summary information for players who played during the match
}
\description{
Gets player data for a match. Recording injuries received and SPP gained.
}
