% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_step.R
\name{plot.bb2_game_step}
\alias{plot.bb2_game_step}
\title{Plot a BB2 game step}
\usage{
\method{plot}{bb2_game_step}(x, ..., alpha = 1)
}
\arguments{
\item{x}{a \code{bb2_game_step} object}

\item{...}{}
}
\description{
Presents a diagnotic plot of the current board state of a \code{bb2_game_step} object. Likely to have been created from a replay file with \link{read_replay_step}.
}
