% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay.R
\name{print.bb2_replay}
\alias{print.bb2_replay}
\title{Print a \code{bb2_replay}}
\usage{
\method{print}{bb2_replay}(x, ...)
}
\arguments{
\item{x}{a \code{bb2_replay} object}

\item{...}{}
}
\description{
Print a \code{bb2_replay}
}
