% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{process_game}
\alias{process_game}
\title{Simulate game}
\usage{
process_game(team_list, weights)
}
\arguments{
\item{team_list}{List of team objects with home team, then away team}

\item{weights}{Vector of weights for probabilities. Length three, order is CCL (TV difference), Home team, Away team}
}
\value{
Return the team list with updated record and points tally.
}
\description{
Simulates game outcome between two teams and updates team's record and points tally.
}
