% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{process_round}
\alias{process_round}
\title{Simulate Round}
\usage{
process_round(rnd, teams, schedule, weights)
}
\arguments{
\item{rnd}{Round number to simulate}

\item{teams}{List of all teams in the division}

\item{schedule}{Full schedule for division. Data frame with headers \code{round}, \code{home_team},\code{away_team}}

\item{weights}{Relative weights for probabilities provided to the \code{calc_game_probs} function}
}
\value{
Returns the team list with all games for the round simulated
}
\description{
Simulates the results of a round of matches for a league.
}
