% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{process_season}
\alias{process_season}
\title{Simulate season}
\usage{
process_season(teams, schedule, weights = list(c(60, 20, 20)),
  simulation_run = 0)
}
\arguments{
\item{teams}{List of teams in the league constructed with \code{\link{make_team}}}

\item{schedule}{Data frame containing the league schedule. Must contain \code{round}, \code{home_team}, and\code{away_team}}

\item{weights}{Vector containing weights for combining probabilities from CCL data and a team's own record. Provided to the \code{\link{calc_game_probs}} function}

\item{simulation_run}{Number of this simulation run. Used to print logging info to console so monitor simulation progress when performed in bulk}
}
\value{
Data frame containing the predicted results of the season.
}
\description{
Simulates an entice season from a list of teams and their schedule
}
