% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay.R
\name{read_replay}
\alias{read_replay}
\title{Read BB2 replay file (.bbrz)}
\usage{
read_replay(replay_file_path, zipped = T)
}
\arguments{
\item{replay_file_path}{path to replay file. Can be relative or absolute path.}
}
\value{
a \code{bb2_replay} object containing relevant metadata as well as the the full replay file xml data for further processing
}
\description{
Reads a .bbrz BB2 replay file and extracts the relevant game metadata.
}
