% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay_step.R
\name{read_replay_step}
\alias{read_replay_step}
\title{Read BB2 replay step}
\usage{
read_replay_step(replay, stepID)
}
\arguments{
\item{replay}{A \code{bb2_replay} object created with read_replay}

\item{stepID}{Integer specifying the index of the replay step to extract}
}
\value{
A \code{bb2_game_step} object
}
\description{
Extracts a single replay step from a replay file
}
