% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_x_bb}
\alias{scale_x_bb}
\title{Position scales for blood bowl pitch}
\usage{
scale_x_bb(...)
}
\arguments{
\item{...}{arguments passed to \code{scale_(x|y)_continuous()}}
}
\description{
Position scales for blood bowl pitch
}
