% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_stats.R
\name{score}
\alias{score}
\title{Get team score}
\usage{
score(match_data, team)
}
\arguments{
\item{match_data}{Match data object from \code{get_game_stats}}

\item{team}{One of \code{"home"} or \code{"away"} to identify the team}
}
\value{
Number of touchdowns scored by the specified team
}
\description{
Get team score
}
