% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{set_up_new_test_file}
\alias{set_up_new_test_file}
\title{Sets up a new test replay file on restart without long data load process}
\usage{
set_up_new_test_file()
}
\description{
Sets up a new test replay file on restart without long data load process
}
