% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{star_player_name}
\alias{star_player_name}
\title{Star Player Names}
\usage{
star_player_name(name)
}
\arguments{
\item{name}{Internal representation of Star Player name}
}
\value{
Human readable Star Player name
}
\description{
Converts internal Star Player names into human readable forms
}
