% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_stats.R
\name{stat_total}
\alias{stat_total}
\title{Total match statistics}
\usage{
stat_total(match_data, stat, team)
}
\arguments{
\item{match_data}{Match data object from \code{get_game_stats}}

\item{stat}{Name of the statistic to collect}

\item{team}{One of \code{"home"} or \code{"away"} to identify the team}
}
\value{
The sum of the individual player stats for the team
}
\description{
Calculate the total match statistics for a team by adding the individual player statistics. Slightly slower than using the pre-summarised stats in .$RowMatch, but these are inaccurate - eg. crowdsurfs aren't recorded at all.
Available statistics are \code{"inflictedCasualties", "inflictedCatches", "inflictedDead", "inflictedInjuries", "inflictedInterceptions", "inflictedKO", "inflictedMetersPassing", "inflictedMetersRunning", "inflictedPasses", "inflictedPushOuts", "inflictedStuns", "inflictedTackles", "inflictedTouchdowns", "sustainedCasualties", "sustainedDead", "sustainedInjuries", "sustainedInterceptions", "sustainedKO", "sustainedStuns", "sustainedTackles"}
}
\examples{
m = get_game_stats(uuid = "1000304d15")
home_kills <- stat_total(m, "inflictedDead", "home")
}
