% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{state_to_casualty}
\alias{state_to_casualty}
\title{Casualty state to effect}
\usage{
state_to_casualty(state)
}
\arguments{
\item{state}{string describing the casualty}
}
\value{
a string describing the effect of the casualty (MNG/Niggle/etc.)
}
\description{
Casualties are stored internally as specific injuries (from CRP table). This converts them into the in game effect on the player.
}
