% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nufflytics.R
\name{theme_nufflytics}
\alias{theme_nufflytics}
\title{Nufflytics theme}
\usage{
theme_nufflytics(...)
}
\arguments{
\item{...}{extra parameters not currently used}
}
\value{
a modified \code{ggplot} theme for use on nufflytics.com
}
\description{
Nufflytics theme
}
