% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{update_probs}
\alias{update_probs}
\title{Update win probabilities}
\usage{
update_probs(team)
}
\arguments{
\item{team}{A team object from \code{make_team}}
}
\value{
Returns the passed team object with win probabilities updated for additional games played
}
\description{
Takes a team object and updates it's internal win probabilities to accurately match the W/D/L record. Should be run once the results of a game simulation have been included in the team's record if required.
}
