% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{update_tv}
\alias{update_tv}
\title{Update TV}
\usage{
update_tv(team)
}
\arguments{
\item{team}{A team object from \code{\link{make_team}}}
}
\value{
Returns the passed team object with an updated TV
}
\description{
Takes a team object and changes it's TV to something new. Takes the TV change from CCL data for the relevant race within +- 20 TV of the team's pre-game TV. To avoid single occurance outliers having a large impact on the results, sample three options and take the median
}
